fix: move multiplatform tests to its own file to share logic across skaffold commands
